requirements.txt
Django==3.2.15
django-cors-headers==3.10.1
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.4.0
PyJWT==1.7.1
PyJWT==1.7.1